refactor frontend
- added a api directory where all backend calls will be consolidated.
- created a health_check endpoint in the backend
