+++ lib/dashboard/src/api/health_check.js
@@ -1,6 +1,5 @@
 
-
-export const healthCheck = async (deployment_url) => {
+export const health_check = async (deployment_url) => {
     try {
         const response = await fetch(`${deployment_url}/health_check`);
         if (!response.ok) {
