+++ lib/dashboard/src/components/BlockConfigBar/AddBlockSidebar.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Button, Offcanvas, CardText, Form, ListGroup } from 'react-bootstrap';
 import { Tooltip, IconButton, Button as MuiButton } from '@mui/material';
 import { MdDelete } from "react-icons/md";
+import { get_block_code, delete_block } from '../../api/blocks';
 
 // TODO: Move this to a separate file
 const AddBlockSidebar = ({ show, handleClose, sideBarData, handleSelectNode }) => {
@@ -32,38 +33,6 @@ const AddBlockSidebar = ({ show, handleClose, sideBarData, handleSelectNode }) =
 }
 
 export default AddBlockSidebar;
-
-const get_block_code = async (core_block_type, process_type) => {
-  try {
-      const response = await fetch(`http://localhost:8000/get_block_codes?core_block_type=${core_block_type}&process_type=${process_type}`);
-      const data = await response.json();
-      return data
-  } catch (error) {
-      console.error('Error fetching block codes:', error);
-      return [];
-  }
-};
-
-const delete_block = async (blockMetadata, vendorName, blockName) => {
-  try {
-      const response = await fetch(`http://localhost:8000/delete_block_type`, {
-          method: 'DELETE',
-          headers: {
-              'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            block_metadata: blockMetadata,
-            vendor_name: vendorName,
-            display_name: blockName
-          })
-      });
-      const data = await response.json();
-      return data
-  } catch (error) {
-      console.error('Error deleting block:', error);
-      return [];
-  }
-}
 
 
 export const AddBlockComponent = ({blocksData, handleSelectNode}) => {
