+++ lib/dashboard/src/components/BlockConfigBar/Code.js
@@ -6,6 +6,7 @@ import { sha1 } from 'js-sha1';
 import { useAuth } from '../../contexts/AuthContext';
 import { styled } from '@mui/material/styles';
 import { Tabs, Tab, Box, TextField, Button } from '@mui/material';
+import { save_block_code, fetch_block_data } from '../../api/blocks';
 
 
 const StyledTabs = styled((props) => (
@@ -119,43 +120,27 @@ export const SaveCodeConfirmation = ({
   const current_file_name = blockDataConfig.source_path.split('/').pop();
   const [finalFileName, setFinalFileName] = useState(current_file_name);
 
-  const handleConfirmSave = () => {
+  const handleConfirmSave = async () => {
     const codeHash = sha1(code);
     const newFilePath = createNewFilePath(blockDataConfig.source_path, finalFileName);
     const core_block_type = blockDataConfig.core_block_type
     const reference_core_block_type = blockDataConfig.reference_core_block_type
 
-    fetch('http://localhost:8000/save_block_code', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-            "Authorization": `Bearer ${token}` 
-        },
-        body: JSON.stringify({
-            source_code: code,
-            source_path: newFilePath,
-            file_name: finalFileName,
-            core_block_type,
-            reference_core_block_type
-        }),
-    })
-    .then(response => {
-        if (response.ok) {
-            response = response.json()
-            console.log('Source code saved successfully.');
-            onDataChange('source_path', newFilePath);
-            onDataChange('source_code', code);
-            onDataChange('source_hash', codeHash);
-            onDataChange('process_type', 'custom');
-            return response
-        } else {
-            console.error('Failed to save source code.');
-        }
-    })
-    .then(response => {
-      const new_core_block_type = response.new_core_block_type
-      onDataChange('core_block_type', new_core_block_type);
-    })
+    const response = await save_block_code(
+      token, 
+      code, 
+      newFilePath, 
+      finalFileName, 
+      core_block_type, 
+      reference_core_block_type
+    );
+    const new_core_block_type = response.new_core_block_type
+    onDataChange('core_block_type', new_core_block_type);
+    onDataChange('source_path', newFilePath);
+    onDataChange('source_code', code);
+    onDataChange('source_hash', codeHash);
+    onDataChange('process_type', 'custom');
+    
     handleCloseConfirmModal(false);
   }
