+++ lib/dashboard/src/components/FormFields/ToolCalling.js
@@ -5,7 +5,7 @@ import { Formik } from 'formik';
 import { AddBlockComponent } from '../BlockConfigBar/AddBlockSidebar';
 import { NodeConfigurationComponent } from '../BlockConfigBar/NodeSidebar';
 import { createRunConfigForNode } from '../Pages/Workflows/Utils/CreateRunConfig';
-import { fetch_block_data } from '../../utils/workflow/get_block_data';
+import { fetch_block_data } from '../../api/blocks';
 import { MdDelete } from 'react-icons/md';
 import { IconButton, Tooltip } from '@mui/material';
