+++ lib/dashboard/src/components/Pages/Chat/Chat.js
@@ -7,8 +7,7 @@ import DropdownButton from './DropUpUtility';
 import { fetchTemplate, fetchChatHistory } from './Utils/fetchUtils';
 import ChatTextInput from '../../InputHandles/ChatTextInput';
 import { UploadFile, UploadFileNames } from '../../InputHandles/UploadFile';
-import { useInputHandlerHooks } from '../../InputHandles/hooks';
-import { useUploadFileNameHooks } from '../../InputHandles/hooks';
+import { useInputHandlerHooks, useUploadFileNameHooks } from '../../../hooks/input_handlers';
 
 
 const ChatComponent = () => {
