+++ lib/dashboard/src/components/Pages/Home/Home.js
@@ -23,7 +23,7 @@ import { MdDelete } from "react-icons/md";
 import { useNavigate } from "react-router-dom";
 import { NewTemplate } from './newTemplate';
 import { useAuth } from '../../../contexts/AuthContext';
-import { fetch_block_data } from '../../../utils/workflow/get_block_data';
+import { fetch_block_data } from '../../../api/blocks';
 import { Position } from "@xyflow/react";
