+++ lib/dashboard/src/components/Pages/RunTemplate/RunTemplate.js
@@ -22,9 +22,9 @@ import { UploadFile, UploadFileNames } from '../../InputHandles/UploadFile';
 import { createInputPayload } from '../../InputHandles/createInputPayload';
 import { prepareInputBlock } from '../../InputHandles/utils';
 import TextInput from '../../InputHandles/TextInput';
-import { useInputHandlerHooks, useLoadingStateHook } from '../../InputHandles/hooks';
-import { useUploadFileNameHooks } from '../../InputHandles/hooks';
-import { healthCheck } from './Utils';
+import { useLoadingStateHook } from '../../../hooks/use_loading';
+import { useInputHandlerHooks, useUploadFileNameHooks } from '../../../hooks/input_handlers';
+import { checkHealth } from '../../../utils/health_check';
 import Run from '../../RunTemplate/Run';
 
 const RunTemplate = () => {
@@ -126,28 +126,6 @@ const RunTemplate = () => {
       });
   };
 
-  /**
-   * Periodically checks the health of the deployment until it is ready.
-   * Essentially here so we know the server is ready to be queried. This
-   * is needed since when loading HF models, given that its a blocking process,
-   * means that the server won't be ready to respond.
-   * @param {string} deployment_url - The URL of the deployment to check.
-   * @returns {Promise<void>}
-   */
-  const checkHealth = async (deployment_url) => {
-    setLoading({loading: true, message: 'Loading Resources...'});
-    while(true){
-      const health_check = await healthCheck(deployment_url);
-      if (health_check) {
-        setLoading({loading: false, message: ''});
-        break
-      }
-      else {
-        await new Promise(resolve => setTimeout(resolve, 2000)); // Wait for 2 seconds
-      }
-    }
-  };
-
     useEffect(() => {
         
         // If url is not available, fetch it from the backend using the id
@@ -189,7 +167,7 @@ const RunTemplate = () => {
                 const inputPayload = createInputPayload(backendTemplate['input'])
                 setInputData((prev) => ({ ...prev, ...inputPayload }));
                 setLoading({loading: false, message: ''});
-                checkHealth(data.deployment_url);
+                checkHealth(data.deployment_url, setLoading);
             })
             .catch(err => {
               setError(err.message);
