+++ lib/dashboard/src/components/Pages/Workflows/TestWorkflow.js
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from 'react'
 import { Button, Modal } from 'react-bootstrap'
 import Run from '../../RunTemplate/Run'
-import { useInputHandlerHooks, useLoadingStateHook, useUploadFileNameHooks } from '../../InputHandles/hooks'
+import { useInputHandlerHooks, useUploadFileNameHooks} from '../../../hooks/input_handlers'
+import { useLoadingStateHook } from '../../../hooks/use_loading'
 import { processNodeDataForBackend } from './Utils/ProcessDataForBackend'
 import { prepareInputBlock } from '../../InputHandles/utils'
 import { createInputPayload } from '../../InputHandles/createInputPayload'
