+++ lib/dashboard/src/components/Pages/Workflows/Workflow.js
@@ -26,8 +26,7 @@ import nodeTypes from '../../Block/BlockTypes';
 
 // Utility function imports
 import { processNodeDataForBackend } from './Utils/ProcessDataForBackend';
-import { FetchTaskRegistry } from '../../../utils/workflow/get_task_registry';
-import { fetch_block_data } from '../../../utils/workflow/get_block_data';
+import { fetch_task_registry, fetch_block_data } from '../../../api/blocks';
 import { 
   checkCollision, 
   calculatePushPosition, 
@@ -268,7 +267,7 @@ export default function Flow() {
 
 
   const addNodeButtonClick = async () => {
-    const blockTypes = await FetchTaskRegistry()
+    const blockTypes = await fetch_task_registry()
     if (blockTypes) {
       setAddNodeSideBarData(blockTypes)
       handleAddNodeSideBarShow()
