+++ lib/dashboard/src/hooks/input_handlers.js
@@ -25,14 +25,3 @@ export const useUploadFileNameHooks = () => {
         setUploadFileName
     }
 }
-
-export const useLoadingStateHook = () => {
-    const [loading, setLoading] = useState({
-        loading: false,
-        message: 'Loading...'
-    });
-    return {
-        loading,
-        setLoading
-    }
-}
