+++ lib/dashboard/src/hooks/use_loading.js
@@ -1,0 +1,12 @@
+import { useState } from "react";
+
+export const useLoadingStateHook = () => {
+    const [loading, setLoading] = useState({
+        loading: false,
+        message: 'Loading...'
+    });
+    return {
+        loading,
+        setLoading
+    }
+}
